Centralize precommit scripts
