+++ package.json
@@ -6,19 +6,20 @@
   "license": "BSD-3-Clause",
   "private": true,
   "scripts": {
-    "_lint-project": "npx eslint --ignore-path .gitignore --ignore-pattern packages/",
-    "lint-project": "npm run _lint-project .",
+    "lint-project": "eslint --ignore-pattern packages/ .",
     "prepare": "husky install"
   },
   "devDependencies": {
     "@eternagame/eslint-plugin": "file:packages/eslint-plugin",
     "@eternagame/jest-utils": "file:packages/jest-utils",
+    "@eternagame/lint-staged-utils": "file:packages/lint-staged-utils",
     "@eternagame/nx-plugin": "file:packages/nx-plugin",
     "@eternagame/nx-spawn": "file:packages/nx-spawn",
     "@eternagame/tsconfig": "file:packages/tsconfig",
     "@eternagame/vite-utils": "file:packages/vite-utils",
     "@nrwl/eslint-plugin-nx": "^14.5.0",
     "@types/jest": "^28.1.6",
+    "@types/micromatch": "^4.0.2",
     "@types/node": "^18.6.2",
     "@types/yargs": "^17.0.10",
     "@typescript-eslint/eslint-plugin": "^5.31.0",
@@ -29,7 +30,6 @@
     "husky": "^8.0.1",
     "jest": "^28.1.3",
     "lint-staged": "^13.0.3",
-    "micromatch": "^4.0.5",
     "node-dev": "^7.4.3",
     "nx": "^14.5.0",
     "shx": "^0.3.4",
