+++ packages/lint-staged-utils/README.md
@@ -1,0 +1,3 @@
+# @eternagame/lint-staged-utils
+
+Eterna standard lint-staged setup and utilities
