+++ packages/lint-staged-utils/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "references": [
+    {
+      "path": "./tsconfig.build.json"
+    },
+    {
+      "path": "./tsconfig.spec.json"
+    }
+  ]
+}
