+++ packages/lint-staged-utils/tsconfig.spec.json
@@ -1,0 +1,16 @@
+{
+  "extends": "@eternagame/tsconfig/tsconfig.jest-node.json",
+  "references": [
+    {
+      "path": "./tsconfig.build.json"
+    }
+  ],
+  "include": [
+    "**/*.spec.ts",
+    "**/*.test.ts",
+    "**/__tests__/**/*.ts",
+    "**/test/**/*.ts",
+    "**/tests/**/*.ts",
+    "**/spec/**/*.ts"
+  ]
+}
