+++ packages/lint-staged-utils/vite.config.ts
@@ -1,0 +1,3 @@
+import getConfig from '@eternagame/vite-utils';
+
+export default getConfig({ type: 'lib', env: 'node' });
