+++ packages/nx-plugin/src/generators/preset/files/lint-staged.config.mjs__tmpl__
@@ -1,0 +1,7 @@
+import lintStagedBase from '@eternagame/lint-staged-utils';
+
+export default function lintStaged(stagedFiles) {
+  return [
+    ...lintStagedBase(stagedFiles),
+  ];
+}
