+++ packages/nx-plugin/src/generators/preset/index.ts
@@ -59,7 +59,6 @@ function addDependencies(tree: Tree) {
         '@typescript-eslint/eslint-plugin',
         'husky',
         'lint-staged',
-        'micromatch',
       ]),
     },
   );
